+++ package.json
@@ -118,7 +118,6 @@
     "karma-typescript": "^4.1.1",
     "lerna": "^3.10.7",
     "mocha": "^7.1.2",
-    "mongodb": "^3.6.3",
     "mongodb-download-url": "^0.6.1",
     "mongodb-js-precommit": "^2.0.0",
     "node-codesign": "^0.3.2",
