--- packages/java-shell/src/test/resources/URI.txt
@@ -1,1 +1,0 @@
-mongodb://localhost:27017
